+++ language-support/java/bindings/src/test/scala/com/daml/ledger/javaapi/data/CommandSpec.scala
@@ -1,0 +1,20 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.javaapi.data
+
+import com.daml.ledger.javaapi.data.Generators._
+import org.scalatest.prop.GeneratorDrivenPropertyChecks
+import org.scalatest.{FlatSpec, Matchers}
+
+class CommandSpec extends FlatSpec with Matchers with GeneratorDrivenPropertyChecks {
+
+  implicit override val generatorDrivenConfig: PropertyCheckConfiguration =
+    PropertyCheckConfiguration(minSize = 1, sizeRange = 3)
+
+  "Command.fromProto" should "convert Protoc-generated instances to data instances" in forAll(
+    commandGen) { cmd =>
+    val converted = Command.fromProtoCommand(cmd)
+    Command.fromProtoCommand(converted.toProtoCommand) shouldEqual converted
+  }
+}